重新生成Proxy TS脚本

+++ AowenPlatform/aspnet-core/src/AowenPlatform.Application/Cars/Dto/CarDto.cs
@@ -7,7 +7,7 @@ using System.Text;
 namespace AowenPlatform.Cars.Dto
 {
     [AutoMapFrom(typeof(Car))]
-    public class CarDto : EntityDto<long>
+    public class CarDto : EntityDto<int>
     {
         /// <summary>
         /// 车型名称
